test-patchbomb: sed -e may not work with extended regexps
Non-GNU versions of the sed(1) utility (i.e. on Solaris or BSD)
may not support extended regexps (i.e. including \| branches),
or may need a special option to enable it (like -E instead of -e).

Expand the sed RE in tests/test-patchbomb to fix running this
test on FreeBSD and Solaris.